+++ Demo/Shared/DemoApp.swift
@@ -10,6 +10,7 @@ import RichTextKit
 
 @main
 struct DemoApp: App {
+
     var body: some Scene {
         WindowGroup {
             ContentView()